+++ src/components/Table/Table.tsx
@@ -268,6 +268,13 @@ interface Props<TData> {
    * @type {DataFrame[]}
    */
   panelData: DataFrame[];
+
+  /**
+   * Highlight Rows On Hover
+   *
+   * @type {boolean}
+   */
+  highlightRowsOnHover?: boolean;
 }
 
 /**
@@ -357,6 +364,7 @@ export const Table = <TData,>({
   replaceVariables,
   stripedRows,
   panelData,
+  highlightRowsOnHover,
 }: Props<TData>) => {
   /**
    * Styles and Theme
@@ -701,6 +709,7 @@ export const Table = <TData,>({
                 onDelete={deleteData.onStart}
                 rowHighlightConfig={rowHighlightConfig}
                 isHighlighted={editableData.row?.id !== row.id && isHighlighted}
+                highlightRowsOnHover={highlightRowsOnHover}
                 panelData={panelData}
               />
             );