+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [3.2.0] - Unreleased
 
-This update brings new features and enhancements designed to improve functionality and user experience.
+### ✨ Added
+
+- Added option to highlight the row on hover. ([#365](https://github.com/VolkovLabs/business-table/pull/365))
 
 ### 🔄 Fixed